Move pages/components into components folder at the root level to fix build issues
