+++ __tests__/components/homePageHeader.test.tsx
@@ -1,4 +1,4 @@
-import HomePageHeader from "@/pages/components/HomePageHeader";
+import HomePageHeader from "@/components/HomePageHeader";
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
