+++ __tests__/skeletonChart.test.tsx
@@ -1,4 +1,4 @@
-import SkeletonChart from "@/pages/components/SkeletonChart";
+import SkeletonChart from "@/components/SkeletonChart";
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 
@@ -6,7 +6,7 @@ const dummySkeletonProps = {
   rootTestId: "someTestId",
   heading: "someHeading",
   subheader: "someSubheader",
-  href: "someHref"
+  href: "someHref",
 };
 
 describe("Skeleton Chart", () => {
@@ -30,29 +30,33 @@ describe("Skeleton Chart", () => {
     expect(screen.getByTestId("someTestIdLink")).toBeInTheDocument;
   });
 
-  it("sets the correct outline image for the massAttendance skeleton", ()=>{
+  it("sets the correct outline image for the massAttendance skeleton", () => {
+    render(
+      <SkeletonChart
+        {...{ ...dummySkeletonProps, rootTestId: "massAttendance" }}
+      />
+    );
 
-        render(<SkeletonChart {...{...dummySkeletonProps, rootTestId:"massAttendance"}} />);
-
-        const outline = screen.getByTestId("massAttendanceSkeletonGraph");
+    const outline = screen.getByTestId("massAttendanceSkeletonGraph");
     expect(outline).toHaveAttribute(
       "src",
       "images/minimalist_National_Mass_Attendance.png"
     );
+  });
 
-  })
+  it("sets the correct outline image for the conversions skeleton", () => {
+    render(
+      <SkeletonChart
+        {...{ ...dummySkeletonProps, rootTestId: "conversions" }}
+      />
+    );
 
-  it("sets the correct outline image for the conversions skeleton", ()=>{
-
-        render(<SkeletonChart {...{...dummySkeletonProps, rootTestId:"conversions"}} />);
-
-        const outline = screen.getByTestId("conversionsSkeletonGraph");
+    const outline = screen.getByTestId("conversionsSkeletonGraph");
     expect(outline).toHaveAttribute(
       "src",
       "images/minimalist_National_Conversions.png"
     );
-
-  })
+  });
 
   test.todo(
     "correct the above test to not give every skeleton the same outline"
