+++ components/DashboardFooter.tsx
