+++ components/NavBar.tsx
